refactor(mcp): split canvas handles out of handles-adapter
handles-adapter.ts crossed the 800-line lint ceiling; canvas entity
resolution and write routing are a coherent unit of their own.
